morning-brief/requirements.txt
requests
anthropic
python-dotenv
yfinance
